Add ability to define libvirt network
This cleans up some of the logic around determining if the user has
a default libvirt network.  Previously, the code looked for a network
called 'default' and failed out prompting the user to edit a file. Now
we allow uses to pass --virtnetwork <some name from virsh net-list> if
they are in a non-standard setup.  Moreover, we now look to see how many
networks are defined and if only 1, we assume to use that one.  If
multiple networks exist, we look for the 'default' network and use that.

imagefactory:
    * getDefaultIP now accepts a hostnet arguement from --virtnetwork
    * Added conditionals to determine the default network

taskbase:
    * added new variable virtnetwork for consumption by imagefactory
      and installer

installer:
    * added --virtnetwork to command line args and pass to getDefaultIP